* Replaced "old lightning" with "greased lightning"
