+++ UnitTest/UT_MsgFormatter.cpp
@@ -4,7 +4,7 @@
 
 #include <gtest/gtest.h>
 // Other files.
-#include "Lightning/GreasedLightning.h"
+#include "Lightning/Lightning.h"
 
 using namespace lightning;
 using namespace std::string_literals;
