+++ applications/profile-lightning.cpp
@@ -3,7 +3,7 @@
 //
 
 
-#include "Lightning/GreasedLightning.h"
+#include "Lightning/Lightning.h"
 #include <cmath>
 #include <sstream>
 #include <iostream>
